requirements.txt
numpy
pandas
tensorflow
pillow
scikit-learn
opencv-python
matplotlib
seaborn